CMakeLists: Add a CMakeLists
Only does what `make all` and `make install` is able to.

Excluded the examples and others in the makefiles that aren't
used by default since this is just for a basic CMakeList

Went with a single file list since it seems that there were a lot
of going in between folders in the makefiles, which was pretty
ugly imo since it was hard to actually tell what line ment what.