Update macOS instructions to include ninja and meson
This adds the `ninja` dependency for in the brew step and the `meson` package for Python in the macOS instructions. They make it is possible to run `make` in the VMAF repository. The `meson` package requires Python 3.5.2 or greater, which is why it is using `pip3`. This is meant to solve #387.