Look for pthreads-win32 when building with MSVC
Note that MSYS64 identifies itself to Meson as 'cygwin', and has a
pthreads-compliant threads implementation.

The sources for pthreads-win32 were once included in this project as a
requirement for MSVC, but never made it into the Meson config and were
removed at some point (I don't know which happened first).